Address comments that were missing in https://commits.webkit.org/269195@main https://bugs.webkit.org/show_bug.cgi?id=263011 rdar://116806970
Reviewed by Alex Christensen and Chris Dumez.

I did not push all changes related to the review of https://github.com/WebKit/WebKit/pull/18887 in https://commits.webkit.org/269195@main.
Adding them here.

* Source/WebKit/WebProcess/GPU/webrtc/MediaRecorderPrivate.cpp:
(WebKit::MediaRecorderPrivate::MediaRecorderPrivate):
* Source/WebKit/WebProcess/GPU/webrtc/MediaRecorderPrivate.h:

Canonical link: https://commits.webkit.org/269207@main
